+++ ashley/src/test/kotlin/ktx/ashley/EnginesSpec.kt
@@ -21,21 +21,40 @@ object EnginesSpec : Spek({
         assertThat(component.y).isEqualTo(0f)
       }
     }
-    describe("creating a component that has no no-arg constructor") {
+    describe("creating a component without a no-arg constructor") {
       @Suppress("UNUSED_PARAMETER")
-      class NoNoArgConstructorComponent(body: String): Component
+      class MissingNoArgConstructorComponent(body: String): Component
 
       it("should throw an exception if the non-pooled engine was unable to create the component") {
         val nonPooledEngine = Engine()
         assertThatExceptionOfType(CreateComponentException::class.java).isThrownBy {
-          nonPooledEngine.create<NoNoArgConstructorComponent>()
+          nonPooledEngine.create<MissingNoArgConstructorComponent>()
         }
       }
 
       it("should throw an exception if the pooled engine was unable to create the component") {
         assertThatExceptionOfType(CreateComponentException::class.java).isThrownBy {
-          engine.create<NoNoArgConstructorComponent>()
+          engine.create<MissingNoArgConstructorComponent>()
         }
+      }
+    }
+    describe("creating a corrupted component that throws an exception") {
+      class CorruptedComponent : Component {
+        init {
+          throw IllegalStateException()
+        }
+      }
+      it("should throw an exception if the non-pooled engine was unable to create the corrupted component") {
+        val nonPooledEngine = Engine()
+        assertThatExceptionOfType(CreateComponentException::class.java).isThrownBy {
+          nonPooledEngine.create<CorruptedComponent>()
+        }.withRootCauseInstanceOf(IllegalStateException::class.java)
+      }
+
+      it("should throw an exception if the pooled engine was unable to create the corrupted component") {
+        assertThatExceptionOfType(CreateComponentException::class.java).isThrownBy {
+          engine.create<CorruptedComponent>()
+        }.withRootCauseInstanceOf(IllegalStateException::class.java)
       }
     }
     describe("creating a component with configuration") {